+++ liam-python-sdk/tests/__init__.py
@@ -1,1 +1,1 @@
-"""XDB Client Tests"""
+"""LIAM Client Tests"""